api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.10.2
annotated-types==0.7.0
anyio==4.8.0
apifairy==1.3.0
apispec==6.8.1
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.8
dataclasses-json==0.6.7
et_xmlfile==2.0.0
exceptiongroup==1.2.2
flake8==7.1.2
Flask==2.2.3
Flask-HTTPAuth==4.4.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.3
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-text-splitters==0.3.6
langchain_ollama==0.1.5
langsmith==0.3.15
Mako==1.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.0.2
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
propcache==0.3.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.26.0
requests-toolbelt==1.0.0
rq==2.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.9
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.7
webargs==8.6.0
Werkzeug==2.2.3
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
